Adjust pulp-open test
