+++ fault_injection_campaign/faulted_pulp_open/Makefile
@@ -11,9 +11,11 @@ clean:
 
 gvsoc:
 	make -C ../.. TARGETS=pulp-open MODULES=$(CURDIR) BUILDDIR=$(FULL_BUILDDIR)/build INSTALLDIR=$(FULL_BUILDDIR)/install $(OPTIONS) build
+	mkdir -p $(FULL_BUILDDIR)/test
+	cp binary $(FULL_BUILDDIR)/test/
 
 test:
-	$(FULL_BUILDDIR)/install/bin/gvsoc --target=pulp-open --work-dir=$(FULL_BUILDDIR)/work --binary=$(FULL_BUILDDIR)/test/test image flash run $(runner_args)
+	$(FULL_BUILDDIR)/install/bin/gvsoc --target=pulp-open --work-dir=$(FULL_BUILDDIR)/work --binary=$(FULL_BUILDDIR)/test/binary image flash run $(runner_args)
 
 campaign:
-	cd .. && ./faulted_pulp_open/complex.py
+	cd .. && ./faulted_pulp_open/fcm.py
