+++ fault_injection_campaign/faulted_pulp_open/binary
